engine/requirements.txt
requests>=2.31.0
openai>=1.0.0
anthropic>=0.18.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
